+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 6/6/2014
+- [CHANGED] Style/IsInversed to Style/IsInverted
+
 ## 5/18/2014
 - [NEW] Initial commit